修改配置，文件优化获取ip 方式

+++ adsl-server/src/main/java/com/github/aidan/adsl/server/socket/MessageHandler2.java
@@ -1,19 +1,16 @@
 package com.github.aidan.adsl.server.socket;
 
 import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson.JSONException;
 import com.github.aidan.adsl.server.bean.ADSL;
-import com.github.aidan.adsl.server.util.JsonFormatUtil;
 import com.github.aidan.adsl.server.util.RemoteShellTool2;
-import com.github.aidan.adsl.server.websocket.WebSocketImpl;
 import com.mysql.cj.util.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-
 import java.io.*;
 import java.net.ConnectException;
 import java.net.Socket;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.concurrent.CopyOnWriteArraySet;
 
 
 public class MessageHandler2 implements Runnable{
@@ -21,11 +18,10 @@ public class MessageHandler2 implements Runnable{
     public Socket socket;
     private RemoteShellTool2 tool;
     private boolean login =false;
-    private static CopyOnWriteArraySet<MessageHandler2> webSocketSet = new CopyOnWriteArraySet<MessageHandler2>();
     public MessageHandler2(Socket socket) {
         this.socket = socket;
     }
-
+    @Autowired
     private ADSL adsl;
 
     @Override
@@ -35,11 +31,10 @@ public class MessageHandler2 implements Runnable{
         PrintWriter out = null;
         try {
             in = new BufferedReader(new InputStreamReader(this.socket.getInputStream(),"utf-8"));
-            out = new PrintWriter(new OutputStreamWriter(this.socket.getOutputStream(),"utf-8"),true);
+            out = new PrintWriter(new OutputStreamWriter(this.socket.getOutputStream(),"utf-8"));
             SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
             String result = "";
             String body = null;
-            if (tool.login()){
                 while (true){
                     body = in.readLine();
                     System.out.println(sdf.format(new Date())+"   :" +body);
@@ -52,7 +47,7 @@ public class MessageHandler2 implements Runnable{
                         case "5" :result = reboot(); break;
                         default:
                          if (!login){
-                            result ="请登录";
+                             result = login(body);
                         }else {
                          result = "输入错误，请输入1 2 3 4 5";
                         }
@@ -62,7 +57,6 @@ public class MessageHandler2 implements Runnable{
                     System.out.println(sdf.format(new Date())+"   :" +result);
                     out.flush();
                 }
-            }
         } catch (ConnectException e) {
             e.printStackTrace();
             out.println("SSH断开重新链接socket");
@@ -87,8 +81,8 @@ public class MessageHandler2 implements Runnable{
     //1
     public String refreshIp(){
         String result ="";
-        if (!isLogin()){
-            return "请登录格式："+JsonFormatUtil.formatJson(JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root")));
+        if (isLogin()) {
+            return "请登录格式：" + JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root"));
         }
         RefreshBatch refreshBatch = new RefreshBatch();
         result = refreshBatch.execute(tool);
@@ -97,11 +91,19 @@ public class MessageHandler2 implements Runnable{
         return result;
 
     }
-//2
+
+    private boolean isLogin() {
+        if (!login){
+            return true;
+        }
+        return false;
+    }
+
+    //2
     public String getIp(){
         String result ="";
-        if (!isLogin()){
-            return "请登录格式："+JsonFormatUtil.formatJson(JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root")));
+        if (isLogin()) {
+            return "请登录格式：" + JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root"));
         }
         try {
             result = tool.exec1("wget -qO- -t1 -T2 ipinfo.io/ip").trim();
@@ -118,8 +120,8 @@ public class MessageHandler2 implements Runnable{
     // 3
     public String getPppoeIp(){
         String result ="";
-        if (!isLogin()){
-            return "请登录格式："+JsonFormatUtil.formatJson(JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root")));
+        if (isLogin()) {
+            return "请登录格式：" + JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root"));
         }
         try {
             result = tool.exec1("ip a | grep ppp* |grep inet | awk '{print $2}'").trim();
@@ -139,8 +141,8 @@ public class MessageHandler2 implements Runnable{
     //4
     public String retryGetIp(){
         String result ="";
-        if (!isLogin()){
-            return "请登录格式："+JsonFormatUtil.formatJson(JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root")));
+        if (isLogin()) {
+            return "请登录格式：" + JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root"));
         }
         try {
             result = tool.exec1("wget -qO- -t1 -T2 ipinfo.io/ip").trim();
@@ -158,6 +160,9 @@ public class MessageHandler2 implements Runnable{
     //5
     public String reboot(){
         String result ="";
+        if (isLogin()) {
+            return "请登录格式：" + JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root"));
+        }
         try {
             result = tool.exec1("reboot").trim();
         } catch (IOException e) {
@@ -167,14 +172,32 @@ public class MessageHandler2 implements Runnable{
         return "重启中4分钟后连接socket";
     }
 
-    public boolean isLogin(){
 
+    public String authenticationFormat(String body){
+        String result ="格式正确";
         try {
-            login = tool.login();
-        } catch (IOException e) {
+            adsl = JSON.parseObject(body,ADSL.class);
+        }catch (JSONException e){
             e.printStackTrace();
+            result = "格式错误,正确格式为："+JSON.toJSONString(new ADSL("127.0.0.1",8080,"q123456","root"));
         }
-        return login;
+        return result;
+    }
+
+    public String login(String body){
+       String result = authenticationFormat(body);
+       if ("格式正确".equals(result)){
+           tool = new RemoteShellTool2(adsl.getHost(), adsl.getPort(), adsl.getUsername(),
+                   adsl.getPwd(), "utf-8");
+           try {
+               login=tool.login();
+               result = "登陆成功";
+           } catch (IOException e) {
+               e.printStackTrace();
+               result ="登陆失败,登陆信息不匹配";
+           }
+       }
+       return result;
     }
 
 }
